requirements.txt
asgiref==3.5.2
channels==4.0.0
constantly==15.1.0
cryptography==38.0.3
daphne==4.0.0
Django==4.1.3
hyperlink==21.0.0
django-resized==1.0.2
Pillow==9.2.0
django-cors-headers==3.13.0
djangorestframework~=3.14.0
